release/version-1.1
New on this release:

Due to ChatGPT's last interface update, the extension didn't work anymore in some areas of the website. Now that's fixed.

Complete overhaul of the extension's interface.

The colors of the color themes were updated.

Now the page reloads automatically when a theme is applied.

The extension wasn't working properly if the user used ChatGPT in Light Mode. Now that's fixed.